wip : vb2.0 wip
