+++ package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "@100mslive/hms-video-react": "^0.2.48",
     "@100mslive/hms-video": "^0.0.76",
-    "@100mslive/hms-virtual-background": "^1.2.1",
+    "@100mslive/hms-virtual-background": "^1.2.2",
     "@craco/craco": "^6.1.1",
     "@material-ui/core": "^4.11.3",
     "@testing-library/jest-dom": "^5.11.4",
